simplified main by using traits and lib
Moved majority of functionality out of main and into functions within
lib.rs. Created a trait with default behaviors for drastically
simplifying testing. No complicated wiremock tests. Need more tests
for branches and possible test integration test for how user would call
program on the cli.
