+++ Cargo.toml
@@ -11,4 +11,5 @@ serde = {version = "1.0", features = ["derive"]}
 chrono = {version = "0.4"}
 clap = {version = "3", features = ["derive"]}
 reqwest = { version = "0.11", features = ["blocking", "json"] }
-serde_json = { version = "1"}
+serde_json = { version = "1" }
+anyhow = { version = "1" }
